+++ agenda-digital-dao/src/main/java/com/everis/academia/java/agenda/digital/dao/ITelefoneDAO.java
@@ -4,5 +4,5 @@ import com.everis.academia.java.agenda.digital.entidades.Telefone;
 
 public interface ITelefoneDAO extends IGenericDAO<Telefone> {
 
-	Boolean delete(Short codigo);
+	Boolean delete(Integer codigo);
 }
